+++ changelog.md
@@ -21,7 +21,7 @@
 
 ## 0.9.0
 
-Released on `?`.
+Released on `2025-12-14`.
 
 - `files`
     + `file` - new module for common file operations
@@ -34,7 +34,7 @@ Released on `?`.
         * `fitsToPandas()` - reading a generic light curves [FITS](https://en.wikipedia.org/wiki/FITS) file into a Pandas table
         * `lightCurveTessToPandas()` - converting a TESS light curve object to a Pandas table
     + `simbad`
-        * `getObjectID()` - fallback with `LIKE` in the `WHERE` clause for problematic identifiers (*such as `SZ  66`*) in `main_id` field of the `basic` table in SIMBAD/CDS database. The problem has been reported to cds-question@unistra.fr, but there was no response yet.
+        * `getObjectID()` - fallback with `LIKE` in the `WHERE` clause for problematic identifiers (*such as `SZ  66`*) in `main_id` field of the `basic` table in SIMBAD/CDS database (*more details in the [documentation](https://phab.decovar.dev/utils/databases/simbad.html#getObjectID)*). The problem has been reported to cds-question@unistra.fr, but there was no response yet.
 
 ## 0.8.2